requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.4
idna==2.8
Pillow==6.1.0
pytz==2019.2
random2==1.0.1
redis==3.3.8
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
